==> Codes/3.5 determinar input shape.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import tensorflow as tf
from keras.backend.tensorflow_backend import set_session
import keras.backend as K

config = tf.ConfigProto()
config.gpu_options.allow_growth = True
config.allow_soft_placement=True
set_session(tf.Session(config=config))

from model.metrics import get_model_memory_usage

from model.MultiResUNet3D import MultiResUnet3D
model = MultiResUnet3D((64, 64, 64, 1), 2)

get_model_memory_usage(model=model, batch_size=1)

from model.u_net import get_3D_unet_soft
model = get_3D_unet_soft((64, 64, 64, 1), 2)

get_model_memory_usage(model=model, batch_size=10)

from model.VNet import vnet
model = vnet((64, 64, 64, 1), 2)

get_model_memory_usage(model=model, batch_size=10)
