==> Codes/4. train test split.ipynb <==
import pandas as pd
from pydicom import dcmread

bboxes = pd.read_csv('./csvs/bboxes.csv').drop(columns='Unnamed: 0')
df = pd.read_csv('./csvs/nodules.csv')

bboxes['length_x_px'] = bboxes.xf - bboxes.xi
bboxes['length_y_px'] = bboxes.yf - bboxes.yi
bboxes['length_z_px'] = bboxes.zf - bboxes.zi

bboxes.head()

paths = df.path.unique()

np.random.shuffle(paths)

splited = np.array(np.split(paths, [int(len(paths) * 0.8)]))

np.save("splited.npy", splited)
