==> Codes/10. generate augmented mask.ipynb <==
from scipy.ndimage import rotate
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from glob import glob
import os

paths = glob('./sin_aumentar/*/*.npz')

path = paths[0]

count = 0
for path in paths:
    file_name = os.path.basename(path)
    new_file_name = path.replace('sin_aumentar', 'augmented')
    new_file_name = new_file_name.replace(file_name, '{}_'+file_name)
    file = np.load(path)
    vol = file['vol']
    mask = file['mask']
    vol_copy = vol.copy()
    mask_copy = mask.copy()
    for j in range(3):
        for i in range(4):
            np.savez_compressed(new_file_name.format(count), vol = vol, mask = mask)
            count += 1
            temp_vol = np.flip(vol, 2)
            temp_mask = np.flip(mask, 2)
            np.savez_compressed(new_file_name.format(count), vol = temp_vol, mask = temp_mask)
            count += 1
            temp_vol = np.flip(temp_vol, 0)
            temp_mask = np.flip(temp_mask, 0)
            np.savez_compressed(new_file_name.format(count), vol = temp_vol, mask = temp_mask)
            count += 1
            temp_vol = np.flip(temp_vol, 2)
            temp_mask = np.flip(temp_mask, 2)
            np.savez_compressed(new_file_name.format(count), vol = temp_vol, mask = temp_mask)
            count += 1
            vol = rotate(temp_vol, 90, axes=(1, 2), mode='nearest')
            mask = rotate(temp_mask, 90, axes=(1, 2), mode='nearest')
        vol = rotate(vol_copy, 90, axes=(0, j + 1), mode='nearest') if j < 2 else None
        mask = rotate(mask_copy, 90, axes=(0, j + 1), mode='nearest') if j < 2 else None

subprocess.call("shutdown -s")
