==> Codes/17.1 augmentation - transformation.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

from sklearn.model_selection import train_test_split
import pandas as pd
import numpy as np

import os
from glob import glob
from tqdm import tqdm_notebook

from scipy.ndimage import rotate, zoom

from mpl_toolkits.axes_grid1 import ImageGrid
from matplotlib import pyplot as plt

def images_grid(images, grid_size):
    fig = plt.figure(figsize=(10, 10))
    grid = ImageGrid(fig, 111, nrows_ncols=grid_size, axes_pad=0.1)

    for ax, im in zip(grid, images):
        ax.imshow(im)

    plt.show()

vol_size = np.array([24, 24, 24])

df =pd.read_csv("/content/drive/MyDrive/LIDC/final_data.csv")

df

agg_func = {'calcification': 'mean', 'internalStructure': 'mean', 'lobulation': 'mean', 'malignancy': 'mean',
            'margin': 'mean', 'path': 'first', 'sphericity': 'mean', 'spiculation': 'mean', 'subtlety': 'mean', 'texture': 'mean',
            'xf': 'first', 'xi': 'first', 'yf': 'first', 'yi': 'first', 'zf': 'first', 'zi': 'first'}

df = df.groupby("final_id").agg(agg_func)

df[['calcification', 'internalStructure', 'lobulation', 'malignancy', 'margin', 'sphericity', 'spiculation', 'subtlety', 'texture']] = df[['calcification', 'internalStructure', 'lobulation', 'malignancy', 'margin', 'sphericity', 'spiculation', 'subtlety', 'texture']]/df[['calcification', 'internalStructure', 'lobulation', 'malignancy', 'margin', 'sphericity', 'spiculation', 'subtlety', 'texture']].max()

df = df[['path', 'calcification', 'internalStructure', 'lobulation', 'malignancy',
       'margin', 'sphericity', 'spiculation', 'subtlety', 'texture']]

len(df)

X_train, X_test, y_train, y_test = train_test_split(df.index.tolist(), df, stratify=df.malignancy, test_size=0.2, random_state=100)

ids_train = X_train
ids_test = X_test

tags_train = (df.loc[ids_train].malignancy.values > 0.5).astype(np.int)
tags_test = (df.loc[ids_test].malignancy.values > 0.5).astype(np.int)

pos_train = tags_train.sum() # pos train
pos_train

pos_test = tags_test.sum() # pos train
pos_test

neg_train = len(tags_train) - pos_train # neg train
neg_train

neg_test = len(tags_test) - pos_test # neg test
neg_test

pos_train + pos_test + neg_train + neg_test

pos_train + pos_test + neg_train + neg_test

needed_train = (pos_train - neg_train) * 36
needed_train

needed_test = (pos_test - neg_test)*36
needed_test

total_needed = needed_train + needed_test
total_needed

# cantidad de parches luego de preprocesamiento
((pos_train + neg_train) * 36) + needed_train

files = {file.path: file.stat().st_size for file in os.scandir('./binary/')}
files = list(files.items())
files = sorted(files, key = lambda x: x[1], reverse=True)
files = [i[0] for i in files]

save_path = './nodules_augmented/{}/c_{}_{}_{}'

count = 0

for ii, file_path in enumerate(files):
    if count >= total_needed:
        break
    
    file = np.load(file_path)
    patch = file['patch']
    tag = file['tag'][0]
    file.close()
    
    if tag == 1:
        continue
        
    subset = "train" if count < needed_train else "test"
    
    current_shape = patch.shape
    patch = zoom(patch, vol_size/current_shape, order=0)
    
    path = save_path.format(subset, 0, 0, count)
    np.savez_compressed(path, patch=patch)
    
    count += 1

print(str(ii)+"/"+str(len(files)))

save_path = './nodules_augmented/{}/{}_{}_{}_{}'

angles = np.arange(10, 90, 10)

angles

reshape = False
mode = "mirror"

for row in tqdm_notebook(df.iterrows(), total = len(df)):
    index = row[0]
    score = row[1][4]
    
    file = np.load('./nodules/{}.npz'.format(index))
    original_patch = file['patch']
    file.close()
    
    current_shape = original_patch.shape
    original_patch = zoom(original_patch, vol_size/current_shape, order=0)
    
    subset = "train" if index in ids_train else "test"
    
    i = 0
    
    path = save_path.format(subset, "o", index, int(score > 0.5), i)
    np.savez_compressed(path, patch=original_patch)
    
    for angle in angles:
        i += 1
        patch = rotate(original_patch, angle=angle, axes=(1, 2), reshape=reshape, mode=mode)
        path = save_path.format(subset, "a", index, int(score > 0.5), i)
        np.savez_compressed(path, patch=patch)
    
    for axis in [0, 1, 2]:
        i += 1
        patch = np.flip(original_patch, axis=axis)
        path = save_path.format(subset, "a", index, int(score > 0.5), i)
        np.savez_compressed(path, patch=patch)

        for angle in angles:
            i += 1
            patch = rotate(patch, angle=angle, axes=(1, 2), reshape=reshape, mode=mode)
            path = save_path.format(subset, "a", index, int(score > 0.5), i)
            np.savez_compressed(path, patch=patch)
